+++ Questionable/Windows/JournalComponents/AlliedSocietyJournalComponent.cs
@@ -25,22 +25,10 @@ internal sealed class AlliedSocietyJournalComponent
     QuestRegistry questRegistry,
     QuestJournalUtils questJournalUtils,
     QuestTooltipComponent questTooltipComponent,
+    Configuration configuration,
     IDalamudPluginInterface pluginInterface,
     UiUtils uiUtils)
 {
-    private static readonly string[] RankNames =
-        ["Neutral", "Recognized", "Friendly", "Trusted", "Respected", "Honored", "Sworn", "Allied"];
-    private readonly AlliedSocietyQuestFunctions _alliedSocietyQuestFunctions = alliedSocietyQuestFunctions;
-    private readonly QuestController _questController = questController;
-    private readonly QuestData _questData = questData;
-
-    private readonly QuestFunctions _questFunctions = questFunctions;
-    private readonly QuestJournalUtils _questJournalUtils = questJournalUtils;
-    private readonly QuestRegistry _questRegistry = questRegistry;
-    private readonly QuestTooltipComponent _questTooltipComponent = questTooltipComponent;
-    private readonly UiUtils _uiUtils = uiUtils;
-    private readonly IDalamudPluginInterface _pluginInterface = pluginInterface;
-
     public void DrawAlliedSocietyQuests()
     {
         using ImRaii.TabItemDisposable tab = ImRaii.TabItem("Allied Societies");
@@ -50,6 +38,35 @@ internal sealed class AlliedSocietyJournalComponent
 #if DEBUG
         if (ImGuiComponents.IconButtonWithText(FontAwesomeIcon.Plus, "Add"))
             addPending = true;
+        if (ImGui.IsItemHovered())
+            ImGui.SetTooltip("Add unchecked quests (from expanded categories) to prio");
+        ImGui.SameLine();
+
+        bool preventQuestCompletion = configuration.Advanced.PreventQuestCompletion;
+        bool abandonQuestBeforeCompletion = configuration.Advanced.AbandonQuestBeforeCompletion;
+        if (ImGuiComponents.IconButton(FontAwesomeIcon.Stop, preventQuestCompletion ? ImGuiColors.DalamudOrange : null))
+        {
+            configuration.Advanced.PreventQuestCompletion = !preventQuestCompletion;
+            pluginInterface.SavePluginConfig(configuration);
+        }
+        if (ImGui.IsItemHovered())
+            ImGui.SetTooltip("Prevent quest completion");
+        if (preventQuestCompletion)
+        {
+            ImGui.SameLine();
+            if (ImGuiComponents.IconButton(FontAwesomeIcon.Trash, abandonQuestBeforeCompletion ? ImGuiColors.DalamudOrange : null))
+            {
+                configuration.Advanced.AbandonQuestBeforeCompletion = !abandonQuestBeforeCompletion;
+                pluginInterface.SavePluginConfig(configuration);
+            }
+            if (ImGui.IsItemHovered())
+                ImGui.SetTooltip("Abandon quest before completion");
+        }
+        else if (abandonQuestBeforeCompletion)
+        {
+            configuration.Advanced.AbandonQuestBeforeCompletion = false;
+            pluginInterface.SavePluginConfig(configuration);
+        }
         ImGui.SameLine();
 #endif
 
@@ -61,8 +78,8 @@ internal sealed class AlliedSocietyJournalComponent
 
         foreach (EAlliedSociety alliedSociety in Enum.GetValues<EAlliedSociety>().Where(x => x != EAlliedSociety.None))
         {
-            List<IQuestInfo> quests = _alliedSocietyQuestFunctions.GetAvailableAlliedSocietyQuests(alliedSociety)
-                .Select(x => _questData.GetQuestInfo(x))
+            List<IQuestInfo> quests = alliedSocietyQuestFunctions.GetAvailableAlliedSocietyQuests(alliedSociety)
+                .Select(x => questData.GetQuestInfo(x))
                 .ToList();
             //if (quests.Count == 0)
             //    continue;
@@ -74,7 +91,7 @@ internal sealed class AlliedSocietyJournalComponent
             {
 #if DEBUG
                 if (quests.Any(x => !x.QuestId.Value.Equals(1569) && ( // Ixal "Deliverance"
-                        !_questRegistry.TryGetQuest(x.QuestId, out Quest? quest) ||
+                        !questRegistry.TryGetQuest(x.QuestId, out Quest? quest) ||
                         (quest.Root.Disabled && !(quest.Root.Comment ?? "").Contains("FATE")) ||
                         (
                             (quest.Root.LastChecked.Date != null && quest.Root.LastChecked.Since(DateTime.Now)!.Value.TotalDays > 90) ||
@@ -89,7 +106,7 @@ internal sealed class AlliedSocietyJournalComponent
                 }
                 else
 #endif
-                if (quests.Any(x => !_questFunctions.IsQuestComplete(x.QuestId)))
+                if (quests.Any(x => !questFunctions.IsQuestComplete(x.QuestId)))
                 {
                     using (ImRaii.PushColor(ImGuiCol.Text, ImGuiColors.DalamudYellow))
                     {
@@ -100,7 +117,7 @@ internal sealed class AlliedSocietyJournalComponent
                     isOpen = ImGui.CollapsingHeader(label);
             }
 
-            _questJournalUtils.ShowQuestGroupContextMenu($"DrawAlliedSocietyQuests{alliedSociety}", quests);
+            questJournalUtils.ShowQuestGroupContextMenu($"DrawAlliedSocietyQuests{alliedSociety}", quests);
 
             if (!isOpen)
                 continue;
@@ -113,8 +130,8 @@ internal sealed class AlliedSocietyJournalComponent
                     if (questsByRank.Count == 0)
                         continue;
 
-                    ImGui.Text(RankNames[i - 1]);
-                    _questJournalUtils.ShowQuestGroupContextMenu($"DrawAlliedSocietyQuests{alliedSociety}/{RankNames[i - 1]}", questsByRank);
+                    ImGui.Text($"{(EAlliedSocietyRank)i}");
+                    questJournalUtils.ShowQuestGroupContextMenu($"DrawAlliedSocietyQuests{alliedSociety}/{(EAlliedSocietyRank)i}", questsByRank);
                     foreach (IQuestInfo quest in questsByRank)
                         DrawQuest((QuestInfo)quest, addPending);
                 }
@@ -129,11 +146,11 @@ internal sealed class AlliedSocietyJournalComponent
 
     private void DrawQuest(QuestInfo questInfo, bool addPending = false)
     {
-        (Vector4 color, FontAwesomeIcon icon, string tooltipText) = _uiUtils.GetQuestStyle(questInfo.QuestId);
+        (Vector4 color, FontAwesomeIcon icon, string tooltipText) = uiUtils.GetQuestStyle(questInfo.QuestId);
         Quest? quest;
         bool fate = false;
         string lastChecked = "";
-        if (!_questRegistry.TryGetQuest(questInfo.QuestId, out quest))
+        if (!questRegistry.TryGetQuest(questInfo.QuestId, out quest))
             color = ImGuiColors.DalamudGrey;
         else
         {
@@ -159,17 +176,17 @@ internal sealed class AlliedSocietyJournalComponent
         string checklistItem = $"{questInfo.Name} ({tooltipText}) {lastChecked}";
         if (fate)
             checklistItem = "(FATE) " + checklistItem;
-        if (_uiUtils.ChecklistItem(checklistItem, color, icon))
-            _questTooltipComponent.Draw(questInfo);
+        if (uiUtils.ChecklistItem(checklistItem, color, icon))
+            questTooltipComponent.Draw(questInfo);
         if (addPending && (color.Equals(ImGuiColors.DalamudRed) || color.Equals(ImGuiColors.DPSRed)))
-            _questController.PriorityManager.Add(questInfo.QuestId);
+            questController.PriorityManager.Add(questInfo.QuestId);
 
-        _questJournalUtils.ShowContextMenu(questInfo, quest, nameof(AlliedSocietyJournalComponent));
+        questJournalUtils.ShowContextMenu(questInfo, quest, nameof(AlliedSocietyJournalComponent));
 
-        if (quest != null && _questController.PriorityManager.Contains(quest))
+        if (quest != null && questController.PriorityManager.Contains(quest))
         {
             ImGui.SameLine();
-            using (_pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
+            using (pluginInterface.UiBuilder.IconFontFixedWidthHandle.Push())
                 ImGui.TextColored(ImGuiColors.DalamudYellow, FontAwesomeIcon.ExclamationCircle.ToIconString());
             if (ImGui.IsItemHovered())
                 ImGui.SetTooltip("This quest is in Priority Quests.");